requirements.txt
babel==2.17.0
docxcompose==1.4.0
docxtpl==0.20.1
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
python-docx==1.2.0
PyYAML==6.0.2
six==1.17.0
typing_extensions==4.14.1
